requirements.txt
about-time==3.1.1
alabaster==0.7.12
alive-progress==2.4.1
apturl==0.5.2
ase==3.21.1
async-generator==1.10
attrs==19.3.0
Automat==0.8.0
autopep8==2.0.0
Babel==2.11.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
beautify==0.1.4
blessings==1.6
blinker==1.4
Brlapi==0.7.0
cajarename==19.7.15
certifi==2022.12.7
chardet==3.0.4
click==8.1.3
cloud-init==23.1.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
configparser==5.0.2
constantly==15.1.0
cryptography==2.8
cudatools==0.0.1
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.30
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
deja-dup-caja==0.0.6
dgl-cu110==0.6.1
dgl-cu113==0.9.1.post1
dglgo==0.0.2
dill==0.3.6
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
docker-pycreds==0.4.0
docutils==0.19
dominate==2.6.0
duplicity==0.8.12.0
easydict==1.9
entrypoints==0.3
exceptiongroup==1.1.0
fasteners==0.14.1
filelock==3.7.1
folder-color-caja==0.0.86
folder-color-common==0.0.86
fonttools==4.34.4
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
googledrivedownloader==0.4
gpg===1.13.1-unknown
gpustat==0.6.0
grapheme==0.6.0
h11==0.14.0
h5py==3.2.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imageio==2.16.1
imagesize==1.4.1
importlib-metadata==5.1.0
incremental==16.10.1
isodate==0.6.0
isort==5.10.1
Jinja2==3.1.2
joblib==1.0.1
json-tricks==3.13.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
littleutils==0.2.2
llvmlite==0.36.0
lockfile==0.12.2
louis==3.12.0
lxml==4.8.0
Magnus==1.0.3
Mako==1.1.0
MarkupSafe==2.1.1
mate-hud==19.10.0
mate-menu==20.4.1
mate-tweak==20.4.0
matplotlib==3.5.2
monotonic==1.5
more-itertools==4.2.0
natsort==8.2.0
netifaces==0.10.4
networkx==2.5.1
numba==0.53.1
numpy==1.23.1
numpydoc==1.5.0
nvidia-cublas-cu11==2022.4.8
nvidia-cublas-cu117==11.10.1.25
nvidia-cuda-runtime-cu11==2022.4.25
nvidia-cuda-runtime-cu117==11.7.60
nvidia-cudnn-cu11==2022.5.19
nvidia-cudnn-cu116==8.4.0.27
nvidia-htop==1.0.5
nvidia-ml-py3==7.352.0
nvidia-pyindex==1.0.9
nvidia-tensorrt==8.4.1.5
oauthlib==3.1.0
ogb==1.3.5
olefile==0.46
onboard==1.4.1
opencv-python==4.5.2.52
outcome==1.2.0
outdated==0.2.2
packaging==21.3
pandas==1.4.3
paramiko==2.6.0
pathtools==0.1.2
pexpect==4.6.0
pickle-mixin==1.0.2
Pillow==9.0.1
platformdirs==2.5.2
prettytable==0.7.2
promise==2.3
protobuf==3.19.0
psutil==5.9.4
pulsemixer==1.5.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.10.0
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.10.2
pygifsicle==1.0.5
Pygments==2.13.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-louvain==0.15
python-xapp==1.8.1
python-xlib==0.23
pytorch-fid==0.3.0
pytransform3d==1.13.0
pytz==2021.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.4
pyzmq==22.3.0
rdflib==5.0.0
rdkit-pypi==2022.9.2
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==0.24.2
scipy==1.8.1
screen-resolution-extra==0.0.0
seaborn==0.12.2
SecretStorage==2.3.1
selenium==4.8.2
sentry-sdk==1.4.3
service-identity==18.1.0
setproctitle==1.1.10
Shapely==1.7.1
shortuuid==1.0.1
simplejson==3.16.0
six==1.14.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sos==4.3
soupsieve==2.3.1
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spicy==0.16.0
ssh-import-id==5.10
subprocess32==3.5.4
systemd-python==234
tensorboardX==2.1
termcolor==1.1.0
threadpoolctl==2.1.0
tomli==2.0.1
torch==1.12.0+cu113
torch-cluster==1.5.9
torch-geometric==1.7.0
torch-scatter==2.0.6
torch-sparse==0.6.9
torch-spline-conv==1.2.1
torchaudio==0.12.0+cu113
torchfile==0.1.0
torchvision==0.13.0+cu113
tornado==6.1
tqdm==4.29.1
trio==0.22.0
trio-websocket==0.9.2
trt-pose==0.0.1
Twisted==18.9.0
typer==0.7.0
typing_extensions==4.4.0
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.14
virtualenv==20.15.1
visdom==0.1.8.9
wadllib==1.3.3
wandb==0.12.5
wcwidth==0.2.5
websocket-client==1.2.1
wsproto==1.2.0
xkit==0.0.0
yaspin==2.1.0
youtube-dl==2020.3.24
zipp==1.0.0
zope.interface==4.7.1
